ai_university_campus_admin_agent/requirements.txt
google-adk
pydantic
sqlalchemy
pytest
python-dotenv
uvicorn
